Refactor: consolidate all proxy functions to a common class (#208)
## Description

In this PR, we moved all of the functions that need to read or write to
the proxy url configuration to one single class `ZaneProxyClient`. In
this class we changed the API to be more generic, and allow for directly
adding & removing the URL configuration for deployments & services. In
addition to this, we made big changes to the way the URL configuration
is structured in production.

we moved from grouping by url configurations by domain : 
```js
{
  "@id": "domain1.com"
   "match": [{ "host": "domain1.com" }],
  "routes": [
       //... url for service1
       //... url for service2
       // etc.
    ],
  "@id": "domain2.com"
   "match": [{ "host": "domain2.com" }],
  "routes": [
       //... url for service2
       // etc.
    ],
}
```

to putting all urls in the root : 
```js
{
  "@id": "service1-domain1.com"
   "match": [{ "host": "domain1.com", "path": "/service-1/*" }],
  "routes": [
       //... url config
    ],
  "@id": "service2-domain1.com"
   "match": [{ "host": "domain1.com", "path": "/service-2/*" }],
  "routes": [
       //... url config
    ],
  "@id": "service2-domain2.com"
   "match": [{ "host": "domain2.com", "path": "/*" }],
  "routes": [
       //... url config
    ],
}
```

This allowed for fixing 2 very silent bugs : 
- Race conditions when modifying urls in the same domain on different
services, we might delete a URL that we added in another service because
the first service has been archived and deleted all its urls, instead we
just have to delete the entry for one service
- if a deployment fails it will not delete old urls and the next
deployment may not have changes with `remove` urls in them so the next
deployment will not remove the old urls for the service, instead we
filter all urls in the proxy and remove the ones that are not attached
to the service anymore.

### ⚠️ If you modify backend code, be sure to run these commands : 

```bash
cd backend
pnpm run openapi # will generate OpenAPI schema at `/openapi/schema.yaml`
pnpm run freeze # will update the `requirements.txt` file if you added new packages
```

### ⚠️ If you modify frontend code, be sure to run these commands : 

```bash
pnpm run format # format the files using biome
```


## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)