Fix/create default temporal namespace (#210)
## Description

In this PR, we setup the process for creating the default namespaces
when zaneops is installed using `temporalio/admin-tools` for doing so.

This is because on production we had to manually setup temporal on
production, which is not acceptable. All of that needs to be created
automatically when zaneops is installed. I did so in this PR.

> [!NOTE]
I also disabled django tests for now as they seem to be stuck on CI and
I can't figure out the reason yet, sometimes they work, and sometimes
they don't, while they work fine locally.

### ⚠️ If you modify backend code, be sure to run these commands : 

```bash
cd backend
pnpm run openapi # will generate OpenAPI schema at `/openapi/schema.yaml`
pnpm run freeze # will update the `requirements.txt` file if you added new packages
```

### ⚠️ If you modify frontend code, be sure to run these commands : 

```bash
pnpm run format # format the files using biome
```


## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)