🔧  run the dev server at `http://localhost:5678` instead of the `sslip.io` domain (#236)
## Description

This change is only for the development environment, instead of starting
the dev server at `app.127-0-0-1.sslip.io`, we use the port where the
frontend is running. This is because the dev server would cause a full
reload when we deploy a new service which has at least one url.

This is an issue with caddy (and not with vite), documented here
https://github.com/caddyserver/caddy/issues/6420. Basically, for
services which have urls, when we deploy them, we need to update the
configuration of caddy, and when this configuration is updated, caddy
closes all the websocket connections causing all clients to reconnect.
This is not a big issue, but in DEV, vite will issue a full reload when
the connection is closed. This was annoying too me, so I made it so that
the DEV server for vite doesn't use the proxy and for each requests to
`/api` to be proxied via vite to the backend running on another port, as
well as to relax the permisssions on the backend so that sessions
domains & csrf aren't enforced on DEV.

### ⚠️ If you modify backend code, be sure to run these commands : 

```bash
cd backend
pnpm run openapi # will generate OpenAPI schema at `/openapi/schema.yaml`
pnpm run freeze # will update the `requirements.txt` file if you added new packages
```

### ⚠️ If you modify frontend code, be sure to run these commands : 

```bash
pnpm run format # format the files using biome
```


## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)