fix: logs rendering perf by removing tanstack virtual overscan (#296)
## Description

Makes logs rendering more performant by removing the overscan of 25 logs to always load. This way if we have only two logs visible in the viewport with BIG ANSI texts, they won't affect the perf as much as if they were 25 (the default). 

It was causing performance issues because the `AnsiHtml` component would run a parsing of all the 25 big logs (which is not very fast), so to mitigate this, we only allow it to parse the logs visible logs. 

The 25 overscan value was an arbitrary number so that the scrolling would be smoother.

fixes #293