refactor: move all services into one global network (#320)
## Description

In this PR, we refactored all the services to be connected to the `zane` global network. The services are still available to other services using their aliases, but the proxy access them using their alias + their slot (`blue|green`) in the `zane` network. This allows for redeploying zaneops with `make deploy[-with-http]` with zero downtime. 

We also made other changes such as : 
- moving `PATH` healthchecks internally, to use the id of the service to make the healthcheck instead of the public deployment url which can be unavailable or protected, this allowed us to remove the `auth_token` in the payload of the  deployment workflow, no need to bypass the authentication to the deployment
- we also added health checks to the services in the production stack to further prevent downtime when redeploying zane-ops