refactor: Use elastic search for logs (#331)
# Description

This is so that search accross a ton of logs is fast. This required storing all logs in elasticsearch and deleting the SimpleLog models previously used for storing logs. The elastic search integration have been created in its own django app located at backend/search with a SearchClient interfacing with ES.

closes #294

I also made some other changes in this PR :

-  Updated the commit_sha shown on the frontend to also show the version and get its data from the API.
-   Updated the API image to use 2 workers
-   Removed the _proxy/logs endpoint as it is not used at all, if we want to see the logs of the proxy, we can inspect them in the terminal.
-   Made it so that the API & temporal workers wait for elasticsearch and the temporal server to be up before starting