refactor: cleanup containers and volumes more aggressively (#352)
## Description

In the previous PR #349, we only removed dangling and unnamed volumes,
this doesn't really do anything if you deploy & archive multiple
services which leaves tagged but unused images, or volumes that are
created outside of zaneops that are not used. this way we free up as
much space as possible.

Of course, we prevent removing volumes used by services, as for images,
if you put a service to sleep, its image will be cleaned up, but what is
cool with docker services is that, they are self healing, when the
service wakes up, it will pull the image again.

> [!WARNING]
> Please do not delete the sections below

### ⚠️ If you modify backend code, be sure to run these commands : 

```bash
cd backend
pnpm run openapi # will generate OpenAPI schema at `/openapi/schema.yaml`
pnpm run freeze # will update the `requirements.txt` file if you added new packages
```

### ⚠️ If you modify frontend code, be sure to run these commands : 

```bash
pnpm install --frozen-lockfile # Install the packages at the exact version listed in the lockfile
pnpm run format # format the files using biome
```


## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)