fix: Occasional 502 Bad Gateway Error on server (#398)
## Description

This was difficult to figure out. 
### The bug 

- On production, sometimes, deploments would become unavailable after a
certain, returning users `502` errors
- This was because of our load balancing policy which would send request
to the old deployment which was already removed instead of sending the
request to the current active deployment.
- The bug is a combination of the load balancing policy which was
`first` and the `fail_duration` which was of **30 seconds**. In the span
of 30s, if caddy marked our new deployment as unavailable, it would not
make any request to that new deployment and instead only make the
request to the first upstream, which is referencing the old deployment
that have already been cleaned up.
 
### The fix 
- We changed the policy from `first` to `round_robin`, this way caddy
would at least switch between upstreams when sending requests to
deployments
- We removed the `fail_duration` so that caddy doesn't remember
incorrectly the new deployment as unavailable.