refactor: simplify proxy configuration for blue/green deployments (#406)
## Description

This is the last PR for fixing the issue with occasional `502 Bad
Gateway` errors. Instead of using load balancing with the blue and green
deployment while one of them is down, we simply replace the previous
deployment with the new deployment when updating the proxy configuration
for the service in production.

In this PR, there are also some other changes : 
- Fixed a bug that sometimes arises in production with service metrics,
data would become missing after a long time, this is because we would
select the incorrect container for the service. it should only get the
metrics of containers for tasks that should be running
- Changed the health check logic to check for all success status codes,
not only `200 OK`